Change bias vector shape, net input calculation
Each input is accepted by the network as a column vector. Net input
formula is changed accordingly, so that the output produces a
units_size*1 column vector that can be passed to the next layer. To make
sure that the input to the network is correctly formatted, set its shape
as it's read from the source.